Create new file with change, emptied stage
